change delete room with images

+++ controllers/roomController.js
@@ -1,4 +1,5 @@
 const { response } = require("express");
+const Image = require("../models/ImageModel");
 
 const Room = require('../models/RoomModel')
 
@@ -94,6 +95,7 @@ const getRooms = async( req, res = response ) => {
 }
 
 const deleteRoom = async( req, res = response ) => {
+    
 
     const idRoom = req.params.idRoom
 
@@ -114,6 +116,20 @@ const deleteRoom = async( req, res = response ) => {
                 ok: false,
                 message: 'El usuario no puede eliminar esta habitación'
             })
+        }
+
+        const findImagesRoom = await Image.find({ room: idRoom });
+
+        if( !findImagesRoom ){
+            res.status(200).json({
+                ok: false,
+                message: 'No se encontraron imagenes en esta habitación'
+            })
+            return
+        }
+
+        for (const image of findImagesRoom) {
+            await Image.findByIdAndDelete( image._id )
         }
 
         await Room.findByIdAndDelete( idRoom )
